requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
boto3>=1.24.0
pandas>=1.3.5
openai>=0.27.0
python-dotenv>=0.20.0
pydantic>=1.9.0
python-multipart>=0.0.5
gunicorn>=20.1.0
PyPDF2
python-docx
requests>=2.27.0
